backend/pyproject.toml
[project]
name = "prompt-tester"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=0.9.9",
    "langchain-google-genai>=3.0.0",
    "langgraph-cli>=0.1.0",
    "langchain-openai>=1.0.1",
    "langchain-ollama>=1.0.0a1",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "langfuse>=3.5.2",
    "asyncio>=4.0.0",
    "fastapi>=0.120.4",
    "pydantic>=2.7.0",
    "langchain-core>=0.3.0",
    "langchain>=1.0.2",
    "langgraph>=1.0.1",
    "uvicorn>=0.30.0",
    "websockets>=15.0.1",
]



[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
src = ["src"]
lint.select = [
    "E",    # pycodestyle
    "F",    # pyflakes
    "I",    # isort
    "D",    # pydocstyle
    "D401", # First line should be in imperative mood
    "T201",
    "UP",
]
lint.ignore = [
    "UP006",
    "UP007",
    "UP035",
    "D417",
    "E501",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
testpaths = ["src/test"]
pythonpath = ["src"]
asyncio_mode = "auto"
addopts = "-v -m 'not llm'"
markers = [
    "llm: marks tests that make real LLM API calls"
]

[tool.setuptools.packages.find]
where = ["src"]  



